requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Diffusion models
diffusers>=0.25.0
transformers>=4.30.0
accelerate>=0.20.0

# Metrics
lpips>=0.1.4
clip-benchmark>=1.4.0

# Image processing
Pillow>=9.0.0
torchvision>=0.15.0

# Audio (optional)
# stable-audio-tools  # Install separately if needed
soundfile>=0.12.0
librosa>=0.10.0

# Video (optional)
imageio>=2.25.0
imageio-ffmpeg>=0.4.8

# Visualization
matplotlib>=3.5.0

# For CLIP/CLAP evaluation
openai-clip>=1.0.0

# Optional: Video CLIP
# viclip  # Install separately if needed
